단일 코드 서비스 별로 모듈 분리(common, summarizer, quiz)
