+++ common/src/main/java/org/zerock/finedu/common/repository/QuizRepository.java
@@ -1,0 +1,7 @@
+package org.zerock.finedu.common.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.zerock.finedu.common.entity.QuizEntity;
+
+public interface QuizRepository extends JpaRepository<QuizEntity, Long> {
+}
