+++ quiz/src/main/java/org/zerock/finedu/quiz/dto/request/QuizAnswerRequest.java
@@ -1,0 +1,10 @@
+package org.zerock.finedu.quiz.dto.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class QuizAnswerRequest {
+    private String userAnswer;
+}
