+++ quiz/src/main/java/org/zerock/finedu/quiz/dto/response/QuizAnswerResponse.java
@@ -1,0 +1,13 @@
+package org.zerock.finedu.quiz.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+
+@Getter
+@Builder
+public class QuizAnswerResponse {
+    private Long quizId;
+    private boolean isCorrect;
+    private String explanation;
+}
