+++ settings.gradle
@@ -1,1 +1,5 @@
 rootProject.name = 'FinEdu-Backend'
+include 'common'
+include 'summarizer'
+include 'quiz'
+
