--- src/main/java/org/zerock/finedu/finedubackend/FinEduBackendApplication.java
@@ -1,13 +1,0 @@
-package org.zerock.finedu.finedubackend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class FinEduBackendApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(FinEduBackendApplication.class, args);
-	}
-
-}
