--- src/main/java/org/zerock/finedu/finedubackend/config/OpenAIConfig.java
@@ -1,26 +1,0 @@
-package org.zerock.finedu.finedubackend.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class OpenAIConfig {
-    @Value("${openai.api-key}")
-    private String openApiKey;
-
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-
-    public HttpHeaders createHttpHeaders() {
-        HttpHeaders headers = new HttpHeaders();
-        headers.set("Authorization", "Bearer " + openApiKey);
-        headers.setContentType(MediaType.APPLICATION_JSON);
-        return headers;
-    }
-}
