--- src/main/java/org/zerock/finedu/finedubackend/dto/request/OpenAIRequest.java
@@ -1,13 +1,0 @@
-package org.zerock.finedu.finedubackend.dto.request;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class OpenAIRequest {
-    private String prompt;
-    @Builder
-    OpenAIRequest(String prompt) {
-        this.prompt = prompt;
-    }
-}
