--- src/main/java/org/zerock/finedu/finedubackend/dto/request/QuizAnswerRequest.java
@@ -1,10 +1,0 @@
-package org.zerock.finedu.finedubackend.dto.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class QuizAnswerRequest {
-    private String userAnswer;
-}
