--- src/test/java/org/zerock/finedu/finedubackend/FinEduBackendApplicationTests.java
@@ -1,13 +1,0 @@
-package org.zerock.finedu.finedubackend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class FinEduBackendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
